Implement BFS- and Saturation-based transitive closure generation. Saturation-based method still has bugs.
git-svn-id: svn://svn.code.sf.net/p/meddly/code/branches/cjiang-ev@802 14966d26-f10a-4722-9547-d5dcc201c352

+++ src/forests/evmdd_pluslong.cc
@@ -354,7 +354,7 @@ MEDDLY::evmdd_pluslong::evpimdd_iterator::~evpimdd_iterator()
   delete[] acc_evs;
 }
 
-void MEDDLY::evmdd_pluslong::evpimdd_iterator::getValue(int &tv) const
+void MEDDLY::evmdd_pluslong::evpimdd_iterator::getValue(long &tv) const
 {
   MEDDLY_DCASSERT(acc_evs);
   tv = acc_evs[0];
